Move loaders to new folder
