+++ src/index.js
@@ -2,14 +2,14 @@ export * from './three/DebugTilesRenderer.js';
 
 // three.js
 export { TilesRenderer } from './three/TilesRenderer.js';
-export { B3DMLoader } from './three/B3DMLoader.js';
-export { PNTSLoader } from './three/PNTSLoader.js';
-export { I3DMLoader } from './three/I3DMLoader.js';
-export { CMPTLoader } from './three/CMPTLoader.js';
+export { B3DMLoader } from './three/loaders/B3DMLoader.js';
+export { PNTSLoader } from './three/loaders/PNTSLoader.js';
+export { I3DMLoader } from './three/loaders/I3DMLoader.js';
+export { CMPTLoader } from './three/loaders/CMPTLoader.js';
 export { GLTFCesiumRTCExtension } from './three/gltf/GLTFCesiumRTCExtension.js';
 export { GLTFStructuralMetadataExtension } from './three/gltf/GLTFStructuralMetadataExtension.js';
 export { GLTFMeshFeaturesExtension } from './three/gltf/GLTFMeshFeaturesExtension.js';
-export { GLTFExtensionLoader } from './three/GLTFExtensionLoader.js';
+export { GLTFExtensionLoader } from './three/loaders/GLTFExtensionLoader.js';
 export { EllipsoidRegionHelper, EllipsoidRegionLineHelper } from './three/objects/EllipsoidRegionHelper.js';
 export { SphereHelper } from './three/objects/SphereHelper.js';
 export { Ellipsoid } from './three/math/Ellipsoid.js';
