+++ src/three/loaders/B3DMLoader.d.ts
@@ -1,10 +1,10 @@
-import { B3DMBaseResult } from '../base/B3DMLoaderBase';
-import { FeatureTable, BatchTable } from '../utilities/FeatureTable';
+import { B3DMBaseResult } from '../../base/B3DMLoaderBase';
+import { FeatureTable, BatchTable } from '../../utilities/FeatureTable';
 import { LoadingManager, Group } from 'three';
 import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
 
 interface B3DMScene extends Group {
-	
+
 	batchTable : BatchTable;
 	featureTable : FeatureTable;
 
@@ -13,7 +13,7 @@ interface B3DMScene extends Group {
 export interface B3DMResult extends GLTF, B3DMBaseResult {
 
 	scene : B3DMScene;
-	
+
 }
 
 export class B3DMLoader {
