+++ src/three/loaders/B3DMLoader.js
@@ -1,4 +1,4 @@
-import { B3DMLoaderBase } from '../base/B3DMLoaderBase.js';
+import { B3DMLoaderBase } from '../../base/B3DMLoaderBase.js';
 import { DefaultLoadingManager, Matrix4 } from 'three';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
