+++ src/three/loaders/CMPTLoader.d.ts
@@ -1,6 +1,6 @@
-import { B3DMBaseResult } from '../base/B3DMLoaderBase';
-import { I3DMBaseResult } from '../base/I3DMLoaderBase';
-import { PNTSBaseResult } from '../base/PNTSLoaderBase';
+import { B3DMBaseResult } from '../../base/B3DMLoaderBase';
+import { I3DMBaseResult } from '../../base/I3DMLoaderBase';
+import { PNTSBaseResult } from '../../base/PNTSLoaderBase';
 import { Group, LoadingManager } from 'three';
 
 export interface CMPTResult {
@@ -11,7 +11,7 @@ export interface CMPTResult {
 }
 
 export class CMPTLoader {
-	
+
 	constructor( manager : LoadingManager );
 	load( url : String ) : Promise< CMPTResult >;
 	parse( buffer : ArrayBuffer ) : Promise< CMPTResult >;
