+++ src/three/loaders/PNTSLoader.d.ts
@@ -1,11 +1,11 @@
-import { PNTSBaseResult, PNTSLoaderBase } from '../base/PNTSLoaderBase';
-import { FeatureTable } from '../utilities/FeatureTable';
+import { PNTSBaseResult, PNTSLoaderBase } from '../../base/PNTSLoaderBase';
+import { FeatureTable } from '../../utilities/FeatureTable';
 import { Points, LoadingManager } from 'three';
 
 interface PNTSScene extends Points {
 
 	featureTable : FeatureTable;
-	
+
 }
 
 export interface PNTSResult extends PNTSBaseResult {
